# contracts/source/openapi.yaml
openapi: 3.1.0
info:
  title: SourceHarbor API
  version: 0.1.19
  description: API contract for subscriptions, ingest polling, jobs, videos and artifacts.
servers:
  - url: http://127.0.0.1:9000
paths:
...
  /api/v1/subscriptions:
    get:
      operationId: listSubscriptions
      parameters:
        - in: query
          name: platform
          schema:
            $ref: '#/components/schemas/Platform'
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: enabled_only
          schema:
            type: boolean
      responses:
        '200':
          description: Subscription list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
    post:
      operationId: upsertSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionUpsertRequest'
      responses:
        '200':
          description: Upsert result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionUpsertResponse'
  /api/v1/subscriptions/manual-intake:
    post:
      operationId: submitManualSourceIntake
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualSourceIntakeRequest'
      responses:
        '200':
          description: Mixed manual intake results for subscriptions and one-off source items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualSourceIntakeResponse'
...
  /api/v1/subscriptions/{id}:
    delete:
      operationId: removeSubscription
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
  /api/v1/subscriptions/batch-update-category:
    post:
      operationId: batchUpdateSubscriptionCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Batch category update result
          content:
            application/json:
              schema:
                type: object
...
components:
  schemas:
...
    Platform:
      type: string
      description: Source platform key. `youtube` and `bilibili` are strong-supported video lanes; other string values are allowed for generic RSSHub/RSS substrate usage.
      examples: [youtube, bilibili, rsshub, generic]
    SourceType:
      type: string
      description: Source identifier type. Known strong-supported values include `youtube_channel_id` and `bilibili_uid`; generic strings such as `rsshub_route` remain allowed without claiming per-route verification.
      examples: [youtube_channel_id, bilibili_uid, url, rsshub_route]
    SubscriptionSupportTier:
      type: string
      enum: [strong_supported, generic_supported]
    SubscriptionContentProfile:
      type: string
      enum: [video, article]
...
    Subscription:
      type: object
      required:
        [id, platform, source_type, source_value, support_tier, content_profile, rsshub_route, enabled, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        platform:
          $ref: '#/components/schemas/Platform'
        source_type:
          $ref: '#/components/schemas/SourceType'
        source_value:
          type: string
        source_name:
          type: string
        support_tier:
          $ref: '#/components/schemas/SubscriptionSupportTier'
        content_profile:
          $ref: '#/components/schemas/SubscriptionContentProfile'
        adapter_type:
          type: string
          enum: [rsshub_route, rss_generic]
        source_url:
          type: string
          nullable: true
        rsshub_route:
          type: string
        creator_display_name:
          type: string
          nullable: true
        creator_handle:
          type: string
          nullable: true
        source_homepage_url:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        avatar_label:
          type: string
          nullable: true
        thumbnail_url:
          type: string
          nullable: true
        source_universe_label:
          type: string
          nullable: true
        identity_status:
          type: string
          nullable: true
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        priority:
          type: integer
          minimum: 0
          maximum: 100
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SubscriptionUpsertRequest:
      type: object
      required: [platform, source_type, source_value]
      properties:
        platform:
          $ref: '#/components/schemas/Platform'
        source_type:
          $ref: '#/components/schemas/SourceType'
        source_value:
          type: string
        adapter_type:
          type: string
          enum: [rsshub_route, rss_generic]
        source_url:
          type: string
          nullable: true
        rsshub_route:
          type: string
          nullable: true
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        priority:
          type: integer
          minimum: 0
          maximum: 100
        enabled:
          type: boolean
...
    SubscriptionUpsertResponse:
      type: object
      required: [subscription, created]
      properties:
        subscription:
          $ref: '#/components/schemas/Subscription'
        created:
          type: boolean
    ManualSourceIntakeRequest:
      type: object
      required: [raw_input]
      properties:
        raw_input:
          type: string
        category:
          type: string
          enum: [tech, creator, macro, ops, misc]
          default: misc
        tags:
          type: array
          items:
            type: string
        priority:
          type: integer
          minimum: 0
          maximum: 100
          default: 50
        enabled:
          type: boolean
          default: true
    ManualSourceIntakeResult:
      type: object
      required:
        [line_number, raw_input, target_kind, recommended_action, status, message]
      properties:
        line_number:
          type: integer
        raw_input:
          type: string
        target_kind:
          type: string
          enum: [subscription_source, manual_source_item, unsupported]
        recommended_action:
          type: string
          enum: [save_subscription, add_to_today, unsupported]
        applied_action:
          type: string
          nullable: true
          enum: [save_subscription, add_to_today]
        status:
          type: string
          enum: [created, updated, queued, reused, rejected]
        platform:
          type: string
          nullable: true
        source_type:
          type: string
          nullable: true
        source_value:
          type: string
          nullable: true
        source_url:
          type: string
          nullable: true
        rsshub_route:
          type: string
          nullable: true
        adapter_type:
          type: string
          nullable: true
        content_profile:
          type: string
          nullable: true
        support_tier:
          type: string
          nullable: true
        display_name:
          type: string
          nullable: true
        relation_kind:
          type: string
          nullable: true
        matched_subscription_id:
          type: string
          nullable: true
        matched_subscription_name:
          type: string
          nullable: true
        matched_by:
          type: string
          nullable: true
        match_confidence:
          type: string
          nullable: true
        source_universe_label:
          type: string
          nullable: true
        creator_display_name:
          type: string
          nullable: true
        creator_handle:
          type: string
          nullable: true
        thumbnail_url:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        avatar_label:
          type: string
          nullable: true
        message:
          type: string
        subscription_id:
          type: string
          nullable: true
        job_id:
          type: string
          nullable: true
    ManualSourceIntakeResponse:
      type: object
      required:
        [processed_count, created_subscriptions, updated_subscriptions, queued_manual_items, reused_manual_items, rejected_count, results]
      properties:
        processed_count:
          type: integer
        created_subscriptions:
          type: integer
        updated_subscriptions:
          type: integer
        queued_manual_items:
          type: integer
        reused_manual_items:
          type: integer
        rejected_count:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/ManualSourceIntakeResult'
...